pyproject.toml
[project]
name = "manako-benchmark"
version = "0.1.0"
description = "Rigorous vision AI detection benchmark: SN44 vs SAM3 vs Roboflow (mAP@50)"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "Pillow>=10.0",
    "requests>=2.31",
    "plotly>=5.18",
    "pandas>=2.1",
    "jinja2>=3.1",
    "click>=8.1",
    "tqdm>=4.66",
    "pycocotools>=2.0",
    "onnxruntime>=1.17",
    "opencv-python>=4.7",
    "huggingface-hub>=0.19",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
cuda = ["onnxruntime-gpu>=1.17"]
roboflow = ["inference>=0.9"]

[project.scripts]
manako-bench = "manako_benchmark.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
